backend/requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==4.2.23
django-cors-headers==4.4.0
django-crontab==0.7.1
django-extensions==3.2.3
django-filter==24.3
django-nested-admin==4.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
kombu==5.5.4
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
prompt-toolkit==3.0.51
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2025.2
redis==6.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
